LoginViewVC: Implemented SaveUserToBackEnd(), updateUserInBackEnd(), and fetchUserFromBackEnd().
SaveUserToBackEnd() is ran when user logs in via FB. If user was already in parse database then the code updates the parseDB with the user’s current FB details. If user was not in parseDB, then the user gets added to parseDB as a new entry.

P.S. This code needs some cleaning, but at least it’s functional