Added hamburger view
Note: The MapView in MapViewVC takes precedence over the contentView's PanGesture. Essentially, if the user tries to pan sideways it will move the map as opposed to moving the side menu. I'm not sure how to get around that. One option would be to have a button on the MapView that shows the side menu. I think this would require a delegate since the button would be in the MapViewVC but it would be controlling the leftMarginConstraints in Containter1VC